requirements.txt
aenum==3.1.16
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
attrs==25.3.0
bidict==0.23.1
boto3==1.34.34
botocore==1.34.162
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
dataclasses-json==0.6.7
deepgram-sdk==3.11.0
deprecation==2.1.0
fastapi==0.115.12
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.2
idna==3.10
jmespath==1.0.1
marshmallow==3.26.1
multidict==6.4.3
mypy_extensions==1.1.0
numpy==1.26.3
ollama==0.1.9
packaging==25.0
propcache==0.3.1
pycparser==2.22
pydantic==2.6.1
pydantic_core==2.16.2
pydub==0.25.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-engineio==4.12.0
python-socketio==5.11.1
PyYAML==6.0.2
requests==2.31.0
s3transfer==0.10.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.4.6
starlette==0.46.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
wsproto==1.2.0
yarl==1.20.0
